file renaming, removed unused code, documentation update
