+++ CorpusCallosum/shape/subsegment_contour.py
