+++ CorpusCallosum/utils/visualization.py
@@ -213,48 +213,3 @@ def plot_contours(
 
     Path(output_path).parent.mkdir(parents=True, exist_ok=True)
     plt.savefig(output_path, dpi=300, bbox_inches="tight")
-
-
-def plot_midplane(grid_orig: np.ndarray, orig: np.ndarray) -> None:
-    """Create a 3D visualization of grid points in original image space.
-
-    Parameters
-    ----------
-    grid_orig : np.ndarray
-        Grid points in original space, shape (3, N).
-    orig : np.ndarray
-        Original image for dimension reference.
-
-    Notes
-    -----
-    The function:
-    1. Creates a 3D scatter plot of grid points
-    2. Samples every 40th point to avoid overcrowding
-    3. Sets axis limits based on original image dimensions
-    4. Shows the plot interactively
-    """
-    # Create a figure showing grid points in original space
-
-    # Create 3D plot
-    fig = plt.figure(figsize=(10, 10))
-    ax = fig.add_subplot(111, projection="3d")
-
-    # Plot every 10th point to avoid overcrowding
-    sample_idx = np.arange(0, grid_orig.shape[1], 40)
-    ax.scatter(*grid_orig[:3, sample_idx], c="r", alpha=0.1, marker=".")
-
-    # Set labels
-    ax.set_xlabel("X")
-    ax.set_ylabel("Y")
-    ax.set_zlabel("Z")
-    ax.set_title("Grid Points in Original Image Space")
-
-    # Set axis limits to image dimensions
-    ax.set_xlim(0, orig.shape[0])
-    ax.set_ylim(0, orig.shape[1])
-    ax.set_zlim(0, orig.shape[2])
-
-    # Save plot
-    plt.show()
-    # plt.savefig('grid_points.png')
-    # plt.close()
