--- doc/api/CorpusCallosum.registration.rst
@@ -1,9 +1,0 @@
-CorpusCallosum.registration
-============================
-
-.. currentmodule:: CorpusCallosum.registration
-
-.. autosummary::
-    :toctree: generated/
-
-    mapping_helpers
