+++ doc/api/CorpusCallosum.shape.rst
@@ -6,9 +6,9 @@ CorpusCallosum.shape
 .. autosummary::
     :toctree: generated/
 
-    cc_postprocessing
-    cc_mesh
-    cc_metrics
-    cc_thickness
-    cc_subsegment_contour
-    cc_endpoint_heuristic
+    postprocessing
+    mesh
+    metrics
+    thickness
+    subsegment_contour
+    endpoint_heuristic
