+++ doc/api/CorpusCallosum.transforms.rst
@@ -6,5 +6,5 @@ CorpusCallosum.transforms
 .. autosummary::
     :toctree: generated/
 
-    localization_transforms
-    segmentation_transforms
+    localization
+    segmentation
