--- doc/api/CorpusCallosum.visualization.rst
@@ -1,9 +1,0 @@
-CorpusCallosum.visualization
-==============================
-
-.. currentmodule:: CorpusCallosum.visualization
-
-.. autosummary::
-    :toctree: generated/
-
-    visualization
