+++ doc/api/index.rst
@@ -19,12 +19,10 @@ FastSurfer API
     CorpusCallosum.rst
     CorpusCallosum.data.rst
     CorpusCallosum.localization.rst
-    CorpusCallosum.registration.rst
     CorpusCallosum.segmentation.rst
     CorpusCallosum.shape.rst
     CorpusCallosum.transforms.rst
     CorpusCallosum.utils.rst
-    CorpusCallosum.visualization.rst
     HypVINN.rst
     HypVINN.dataloader.rst
     HypVINN.models.rst
