+++ pyproject.toml
@@ -53,13 +53,13 @@ dependencies = [
     'monai>=1.4.0',
     'meshpy>=2025.1.1',
     'pyrr>=0.10.3',
-    'whippersnappy>=1.3.1',
     'pip>=25.0',
 ]
 
 [project.optional-dependencies]
 doc = [
     'furo!=2023.8.17',
+    'whippersnappy>=1.3.1',
     'memory-profiler',
     'myst-parser',
     'numpydoc',
